+++ app.py
@@ -1,11 +1,8 @@
-from typing import Dict, Optional, Type
-
 import pyocle
 from chalice import Chalice, CognitoUserPoolAuthorizer
 from mongoengine import connect
-from pydantic.main import BaseModel
 
-from chalicelib.form import ContactMessageCreationForm, ContactMessageQuery
+from chalicelib.form import ContactMessageCreationForm, ContactMessageQueryParameters
 from chalicelib.model import ContactMessageCollection
 from chalicelib.service import ContactMessageService
 
@@ -14,14 +11,14 @@ app = Chalice(app_name='contact-message-service')
 connection_string = pyocle.config.connection_string(default='')
 connect(host=connection_string)
 
-contact_message_service = ContactMessageService()
+cms = ContactMessageService()
 authorizer = CognitoUserPoolAuthorizer('portfolio-userpool',
                                        provider_arns=[
                                            'arn:aws:cognito-idp:us-east-2:811393626934:userpool/us-east-2_MLclIlI5Y'])
 
 
 @app.route('/mail', methods=['POST'], cors=True)
-@pyocle.error.error_handler
+@pyocle.response.error_handler
 def create_contact_message():
     """
     Endpoint used for create contact messages. This endpoint is open to anonymous users.
@@ -31,13 +28,13 @@ def create_contact_message():
 
     form = pyocle.form.resolve_form(app.current_request.raw_body, ContactMessageCreationForm)
     identity = app.current_request.context['identity']
-    contact_message = contact_message_service.create_with_identity(form, identity)
+    contact_message = cms.create_with_identity(form, identity)
 
     return pyocle.response.created(contact_message)
 
 
 @app.route('/mail/{identifier}', methods=['GET'], cors=True, authorizer=authorizer)
-@pyocle.error.error_handler
+@pyocle.response.error_handler
 def get_single_contact_message(identifier: str):
     """
     Endpoint used to retrieve contact messages.
@@ -46,12 +43,12 @@ def get_single_contact_message(identifier: str):
     :return: The found contact message
     """
 
-    contact_message = contact_message_service.find_one(identifier)
+    contact_message = cms.find_one(identifier)
     return pyocle.response.ok(contact_message)
 
 
 @app.route('/mail', methods=['GET'], cors=True, authorizer=authorizer)
-@pyocle.error.error_handler
+@pyocle.response.error_handler
 def get_multiple_contact_message():
     """
     Endpoint used to retrieve a specific contact message
@@ -59,11 +56,7 @@ def get_multiple_contact_message():
     :return: The found contact messages
     """
 
-    query = resolve_query(app.current_request.query_params, ContactMessageQuery)
-    contact_messages = contact_message_service.find(**query)
+    query_params = pyocle.form.resolve_query_params(app.current_request.query_params, ContactMessageQueryParameters)
+    contact_messages = cms.find_paginated(**query_params.dict(exclude_none=True))
     collection = ContactMessageCollection(contact_messages)
-    return pyocle.response.ok(collection)
-
-
-def resolve_query(params: Optional[Dict[str, str]], model: Type[BaseModel]):
-    return {} if params is None else pyocle.form.resolve_form(params, model).dict(skip_defaults=True)
+    return pyocle.response.ok(collection, query_params)